LayoutTests/platform/ios-simulator/editing/shadow/bold-twice-in-shadow-expected.txt

CONSOLE MESSAGE: line 3: HierarchyRequestError: DOM Exception 3: A Node was inserted somewhere it doesn't belong.
Applying document.execCommand('Bold') to elements having insertion points (<shadow> or <content>) shoud not cause a crash.

To test manually, make a selection from somewhere in "nested before" to somehwere in "nested after", then press Ctrl+B twice. It should not cause a crash.

before host
not editable
after host
